refactor(comments): extract native ConfirmDialog into the design system
Review follow-up on #392: the Portal + Dialog combo from
CommentDeleteDialog is now a reusable design-system component at
components/native/dialogs/ConfirmDialog, mirroring the web
ConfirmDialog API (title, content, cancelText, action slot) with the
native close-X title row. CommentDeleteDialog is a thin consumer that
keeps only the delete mutation wiring.
