+++ packages/design-system/components/native/dialogs/ConfirmDialog/types.ts
@@ -1,0 +1,10 @@
+import { ReactNode } from 'react'
+
+export interface ConfirmDialogProps {
+  visible: boolean
+  onClose: VoidFunction
+  title: string
+  content?: ReactNode
+  cancelText?: string
+  action: ReactNode
+}
