+++ packages/design-system/components/native/dialogs/index.ts
@@ -1,0 +1,2 @@
+export { default as ConfirmDialog } from './ConfirmDialog'
+export type * from './ConfirmDialog/types'
